+++ ydb/core/kqp/opt/rbo/kqp_rewrite_select.cpp
@@ -1550,8 +1550,13 @@ TExprNode::TListType FindSublinks(const TExprNode::TPtr& node) {
     });
 }
 
+bool IsAsTable(const TExprNode::TPtr input) {
+    return TCoParameter::Match(input.Get()) ||
+           (input->IsCallable("ToList") && TCoParameter::Match(input->HeadPtr().Get()));
+}
+
 TExprNode::TPtr RewriteSublinks(TExprNode::TPtr& node, TExprContext& ctx, const TTypeAnnotationContext& typeCtx, const TKqpOptimizeContext& kqpCtx,
-                              ui64& uniqueSourceIdCounter, THashMap<const TExprNode*, TExprNode::TPtr>& translated) {
+                              ui64& uniqueSourceIdCounter, ui64& uniqueColumnIdCounter, THashMap<const TExprNode*, TExprNode::TPtr>& translated) {
 
     auto sublinks = FindSublinks(node);
     YQL_CLOG(TRACE, ProviderKikimr) << "Sublinks size: " << sublinks.size();
@@ -1566,7 +1571,7 @@ TExprNode::TPtr RewriteSublinks(TExprNode::TPtr& node, TExprContext& ctx, const 
         TNodeOnNodeOwnedMap nodeReplacementMap;
         TExprNode::TPtr newNode;
 
-        auto newSubquery = RewriteSelect(sublink->ChildPtr(4), ctx, typeCtx, kqpCtx, uniqueSourceIdCounter, translated, false);
+        auto newSubquery = RewriteSelect(sublink->ChildPtr(4), ctx, typeCtx, kqpCtx, uniqueSourceIdCounter, uniqueColumnIdCounter, translated, false);
         auto sublinkType = sublink->Child(0)->Content();
 
         if (sublinkType == "expr") {
@@ -1600,7 +1605,6 @@ TExprNode::TPtr RewriteSublinks(TExprNode::TPtr& node, TExprContext& ctx, const 
     }
     return node;
 }
-
 } // anonymous namespace
 
 TExprNode::TPtr RewriteTableEffect(const TExprNode::TPtr& node, TExprContext& ctx, const TKqpOptimizeContext& kqpCtx) {
@@ -1728,21 +1732,21 @@ TExprNode::TPtr RewriteTableEffect(const TExprNode::TPtr& node, TExprContext& ct
     }
 }
 
-
 TExprNode::TPtr RewriteSelect(const TExprNode::TPtr& input, TExprContext& ctx, const TTypeAnnotationContext& typeCtx, const TKqpOptimizeContext& kqpCtx,
-                              ui64& uniqueSourceIdCounter, THashMap<const TExprNode*, TExprNode::TPtr>& translated, bool generateRoot) {
+                              ui64& uniqueSourceIdCounter, ui64& uniqueColumnIdCounter, THashMap<const TExprNode*, TExprNode::TPtr>& translated,
+                              bool generateRoot) {
 
     if(translated.contains(input.Get())) {
         return translated.at(input.Get());
     }
     TVector<TString> finalColumnOrder;
-    // Start from beggining for each proccesed select;
-    ui64 uniqueAggColumnId = 0;
+    // Generated column names must be unique across the whole query.
+    ui64& uniqueAggColumnId = uniqueColumnIdCounter;
 
     TExprNode::TPtr node = input;
 
     if (generateRoot) {
-        node = RewriteSublinks(node, ctx, typeCtx, kqpCtx, uniqueSourceIdCounter, translated);
+        node = RewriteSublinks(node, ctx, typeCtx, kqpCtx, uniqueSourceIdCounter, uniqueColumnIdCounter, translated);
     }
 
     auto setItems = GetSetting(node->Head(), "set_items")->TailPtr();
@@ -1798,7 +1802,7 @@ TExprNode::TPtr RewriteSelect(const TExprNode::TPtr& input, TExprContext& ctx, c
                     if (translated.contains(childExpr.Get())) {
                         subquery = translated.at(childExpr.Get());
                     } else {
-                        subquery = RewriteSelect(childExpr, ctx, typeCtx, kqpCtx, uniqueSourceIdCounter, translated, false);
+                        subquery = RewriteSelect(childExpr, ctx, typeCtx, kqpCtx, uniqueSourceIdCounter, uniqueColumnIdCounter, translated, false);
                     }
 
                     // We need to rename all the IUs in the subquery to reflect the new alias
@@ -1806,6 +1810,20 @@ TExprNode::TPtr RewriteSelect(const TExprNode::TPtr& input, TExprContext& ctx, c
                         .Input(subquery)
                         .Alias(alias)
                         .Done().Ptr();
+                }
+                else if (IsAsTable(childExpr)) {
+                    auto param = childExpr->IsCallable("ToList") ? childExpr->HeadPtr() : childExpr;
+
+                    // clang-format off
+                    auto source = Build<TKqpOpEmptySource>(ctx, node->Pos())
+                        .Input(param)
+                    .Done();
+
+                    fromExpr = Build<TKqpOpReplaceAlias>(ctx, node->Pos())
+                        .Input(source)
+                        .Alias(alias)
+                    .Done().Ptr();
+                    // clang-format on
                 }
                 else {
                     Y_ENSURE(false, TStringBuilder() << "Unsupported callable: " << childExpr->Content());