+++ ydb/core/tx/columnshard/engines/reader/trivial_reader/iterator/sys_view/optimizer/source.h
@@ -22,6 +22,8 @@ private:
     }
 
 public:
+    virtual NColumnShard::TInternalPathId GetPathId() const override;
+
     TSourceData(const ui32 sourceIdx, const ui64 tabletId, const std::shared_ptr<const TGranuleMeta>& granule,
         std::vector<NStorageOptimizer::TTaskDescription>&& tasks, const NColumnShard::TSchemeShardLocalPathId& externalPathId,
         NArrow::TSimpleRow&& start, NArrow::TSimpleRow&& finish, const std::shared_ptr<NReader::NCommon::TSpecialReadContext>& context)